Add support for scoring and team lists
Using Peter's scoring library, this now calculates game points from
all the yaml files in the ./league/ directory.

It also has a list of teams.

At the moment, this just symlinks to some files within the
scoring-2014.git repository, which isn't particularly fun.  Ideally
there would be a directory with an __init__.py file in there that
would make it more usable.