Use the same SRComp instances across requests
Requests were taking ~40ms because the SRComp object was instantiated
each time.  By using the same SRComp object across requests, this
drops to ~15ms on my machine.

Since SRComp doesn't contain any special stuff for working out whether
its gone out of date (and it probably shouldn't, as that may just be
too complicated), this code causes the SRComp object to be cached for
a maximum of 5 seconds.